test(administration): extract the sw-admin-menu spec harness
